# openapi.yaml
openapi: 3.0.3
info:
  title: Everroute API
  version: 1.0.0
  description: REST API contract for Everroute backend and mobile integration.
  contact:
    name: Everroute Engineering
    email: engineering@example.com
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging-api.example.com
    description: Staging
...
paths:
...
  /v1/auth/me:
    get:
      tags: [Auth]
      summary: Current authenticated user context
      description: Returns identifiers from the validated JWT (user, role, organization).
      operationId: getAuthMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthMeResponse:
      type: object
      required: [user_id, role, org_id]
      properties:
        user_id:
          type: string
        role:
          type: string
        org_id:
          type: string
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    UnauthorizedError:
      description: Unauthorized request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: unauthorized
            message: Authentication is required.
...